deplace dashbord logic in its component
