+++ package.json
@@ -13,6 +13,7 @@
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^14.0.0",
+    "react-router-dom": "^6.21.3",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
