+++ src/components/About.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const About = () => {
+  return (
+    <div>
+      <h1>À propos de nous</h1>
+      <p>Ceci est la page à propos de notre application.</p>
+    </div>
+  );
+};
+
+export default About;
